Update networkedsystemrepairkit.ps1
Removes shutdown /r command. Not necessary as restart is built into the /force script. On occassion have seen the /force gpupdate switch hang for an extended amount of time and used shutdown /r to force the restart but it is not necessary to include in the script as you may only need that /r seldomly.